[Bulk of fixes and errors in game beta template]
Sorry for Errors, and the chat, may be fixed, another pages may be still bugs, so must me patient please. Sorry for the chat swearing, really  not works at the moment. And messages or swearing is not allowed, im sorry, fixing this, im fixing this template. Note: Its not possible play really, sorry for this, that the api not works at the moment. And the chat is a test, sorry for the test messages and Bad words, in the chat of website, or numbers, etc. May be fixed.